requirements.txt
cffi==0.8.6
cryptography==0.6.1
pyOpenSSL==0.14
pycparser==2.10
python-dateutil==2.2
six==1.8.0
certifi
